Feat/logged in user

+++ src/actions/actions.ts
@@ -1,8 +1,10 @@
 "use server";
 
+import bcrypt from "bcrypt";
 import { revalidatePath } from "next/cache";
+import { redirect } from "next/navigation";
 
-import { signIn } from "@/lib/auth";
+import { auth, signIn, signOut } from "@/lib/auth";
 import prisma from "@/lib/db";
 import { sleep } from "@/lib/utils";
 import { petFormSchema, petIdeaSchema } from "@/lib/validations";
@@ -10,7 +12,7 @@ import { petFormSchema, petIdeaSchema } from "@/lib/validations";
 // --- user actions ---
 export async function logIn(formData: FormData) {
   const authData = Object.fromEntries(formData.entries());
-  console.log(authData, "authdata");
+  // console.log(authData, "authdata");
 
   await signIn("credentials", authData);
   // Validate the data
@@ -18,9 +20,36 @@ export async function logIn(formData: FormData) {
   // If the data is valid, log the user in
 }
 
+export async function logOut() {
+  // Log the user out using the signOut function from the auth library
+  await signOut({ redirectTo: "/" });
+}
+
+export async function signUp(formData: FormData) {
+  const authData = Object.fromEntries(formData.entries());
+  // console.log(authData, "authdata");
+
+  const hashedPassword = await bcrypt.hash(authData.password as string, 10);
+
+  await prisma.user.create({
+    data: {
+      email: authData.email as string,
+      hashedPassword: hashedPassword,
+    },
+  });
+
+  // Log the user using the signIn function from the auth library which is imported from the auth
+  await signIn("credentials", formData);
+}
+
 // --- pet actions ---
 export async function addPet(newPetData: unknown) {
   await sleep(1000);
+
+  const session = await auth();
+  if (!session?.user) {
+    redirect("/login");
+  }
 
   const validatedPet = petFormSchema.safeParse(newPetData);
   if (!validatedPet.success) {
@@ -28,12 +57,19 @@ export async function addPet(newPetData: unknown) {
   }
   try {
     await prisma.pet.create({
-      data: validatedPet.data,
+      data: {
+        ...validatedPet.data,
+        user: {
+          connect: {
+            id: session?.user?.id,
+          },
+        },
+      },
     });
   } catch (error) {
     console.log(error);
 
-    return { message: "error" };
+    return { message: "Could not add pet" };
   }
 
   revalidatePath("/app", "layout");
@@ -42,6 +78,13 @@ export async function addPet(newPetData: unknown) {
 export async function editPet(petId: unknown, updatedPetData: unknown) {
   await sleep(1000);
 
+  // authentification check
+  const session = await auth();
+  if (!session?.user) {
+    redirect("/login");
+  }
+
+  // validation check
   const validatedPet = petFormSchema.safeParse(updatedPetData);
   if (!validatedPet.success) {
     return { message: "Invalid data" };
@@ -50,6 +93,17 @@ export async function editPet(petId: unknown, updatedPetData: unknown) {
   const validatedPetId = petIdeaSchema.safeParse(petId);
   if (!validatedPetId.success) {
     return { message: "Invalid data" };
+  }
+
+  // authorization check
+  const pet = await prisma.pet.findUnique({
+    where: { id: validatedPetId.data },
+  });
+  if (!pet) {
+    return { message: "Pet not found" };
+  }
+  if (pet.userId !== session?.user?.id) {
+    return { message: "Unauthorized" };
   }
 
   try {
@@ -68,11 +122,33 @@ export async function editPet(petId: unknown, updatedPetData: unknown) {
 export async function deletePet(petId: unknown) {
   await sleep(1000);
 
+  const session = await auth();
+  if (!session?.user) {
+    redirect("/login");
+  }
+
   const validatedPetId = petIdeaSchema.safeParse(petId);
+
   if (!validatedPetId.success) {
     return { message: "Invalid data" };
   }
 
+  // authorization check
+  const pet = await prisma.pet.findUnique({
+    where: {
+      id: validatedPetId.data,
+    },
+  });
+
+  if (!pet) {
+    return { message: "Pet not found" };
+  }
+
+  if (pet.userId !== session?.user?.id) {
+    return { message: "Unauthorized" };
+  }
+
+  // database mutation
   try {
     await prisma.pet.delete({
       where: {
@@ -82,5 +158,6 @@ export async function deletePet(petId: unknown) {
   } catch (error) {
     return { message: "Could not delete" };
   }
+
   revalidatePath("/app", "layout");
 }
